Add Statement Adapter for SDN6 (!427)

+++ library/src/main/kotlin/eu/tib/orkg/prototype/statements/api/statement-queries.kt
@@ -30,7 +30,7 @@ interface RetrieveStatementUseCase {
     ): Iterable<StatementRepresentation>
 
     fun totalNumberOfStatements(): Long
-    fun countStatements(paperId: String): Int
+    fun countStatements(paperId: String): Long
     fun findAllByPredicateAndLabel(
         predicateId: PredicateId,
         literal: String,
